--- .prettierignore
@@ -1,1 +1,0 @@
-webapp/main/routes.js
