+++ webapp/app/landing/Introduction/introduction.less
@@ -2,6 +2,7 @@
 
 .home-introduction {
   position: relative;
+  height: 57vh;
 }
 
 .home-introduction__about-fra {
