+++ webapp/app/landing/KeyFinidings/keyFindings.less
@@ -1,11 +1,12 @@
 @import (reference) 'webapp/main/styles/config';
 
 .home-key-findings {
-  background-color: @ui-bg;
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   align-items: center;
   justify-items: center;
+  background-color: @ui-accent;
+  color: white;
 }
 
 .home-key-findings__map {
