+++ webapp/app/landing/landing.js
@@ -1,18 +1,18 @@
-import './landing.less'
 import React from 'react'
 
+import CountrySelection from '@webapp/components/countrySelection'
 import Partners from '@webapp/components/Partners'
+
 import Introduction from './Introduction'
 import KeyFinidings from './KeyFinidings'
 
-const Landing = () => {
-  return (
-    <div className="home">
-      <Introduction />
-      <KeyFinidings />
-      <Partners />
-    </div>
-  )
-}
+const Landing = () => (
+  <>
+    <CountrySelection />
+    <Introduction />
+    <KeyFinidings />
+    <Partners />
+  </>
+)
 
 export default Landing
