--- webapp/app/landing/landing.less
@@ -1,4 +1,0 @@
-.home {
-  display: grid;
-  grid-template-rows: 57vh repeat(2, auto);
-}
