+++ webapp/components/Header/header.less
@@ -1,6 +1,7 @@
 @import (reference) 'webapp/main/styles/config';
 
 .app-header {
+  height: @ui-app-header-height;
   border-bottom: @ui-thin-border;
   display: grid;
   grid-template-columns: calc(@ui-app-navigation-width - 21px) 1px auto auto;
