+++ webapp/components/countrySelection/components/countryListDownload.js
@@ -1,5 +1,4 @@
 import './countryListDownload.less'
-
 import React from 'react'
 import { useSelector } from 'react-redux'
 
@@ -15,7 +14,6 @@ import useI18n from '@webapp/components/hooks/useI18n'
 import * as CountryState from '@webapp/app/country/countryState'
 
 const CountryListDownload = () => {
-
   const userInfo = useUserInfo()
   if (!isAdministrator(userInfo)) {
     return null
@@ -24,11 +22,11 @@ const CountryListDownload = () => {
   const i18n = useI18n()
   const countries = useSelector(CountryState.getCountriesList)
 
-  const data = countries.map(country => ({
+  const data = countries.map((country) => ({
     name: Country.getListName(i18n.language)(country),
     status: i18n.t(`assessment.status.${Country.getFra2020Assessment(country)}.label`),
     edited: getRelativeDate(Country.getLastEdit(country), i18n) || i18n.t('audit.notStarted'),
-    deskStudy: i18n.t(`yesNoTextSelect.${Country.isFra2020DeskStudy(country) ? 'yes' : 'no'}`)
+    deskStudy: i18n.t(`yesNoTextSelect.${Country.isFra2020DeskStudy(country) ? 'yes' : 'no'}`),
   }))
 
   const headers = [
@@ -40,13 +38,9 @@ const CountryListDownload = () => {
 
   return (
     <div className="country-selection-list__download">
-      <CSVLink
-        className="btn-s btn-primary"
-        target="_blank"
-        filename="FRA-Countries.csv"
-        data={data}
-        headers={headers}>
-        <Icon className="icon-sub icon-white" name="hit-down"/>CSV
+      <CSVLink className="btn-s btn-primary" target="_blank" filename="FRA-Countries.csv" data={data} headers={headers}>
+        <Icon className="icon-sub icon-white" name="hit-down" />
+        CSV
       </CSVLink>
     </div>
   )
