+++ webapp/components/countrySelection/components/countrySelection.less
@@ -8,6 +8,7 @@
   background-color: @ui-bg;
   border-bottom: @ui-thin-border;
   padding: 0 16px;
+  height: 50px;
 }
 
 .btn-toggle-navigation {
