+++ webapp/components/footer/components/sendFeedback.js
@@ -13,8 +13,12 @@ const SendFeedback = () => {
 ${newLine}${newLine}
 *** ${i18n.t('navigation.support.dontDelete')} ***
 ${newLine}${newLine}
-${i18n.t('navigation.support.user')}: ${userInfo.name} (${userInfo.id})
-${newLine}
+${
+  userInfo
+    ? `${i18n.t('navigation.support.user')}: ${userInfo.name} (${userInfo.id})
+        ${newLine}`
+    : ''
+}
 ${i18n.t('navigation.support.platformVersion')}: ${__APPLICATION_VERSION__}
 ${newLine}
 ${i18n.t('navigation.support.userAgent')}: ${navigator.userAgent}
