+++ webapp/main/styles/style.less
@@ -32,6 +32,11 @@ body {
   height: 100%;
 }
 
+#main {
+  height: 100vh;
+  width: 100vw;
+}
+
 h1,
 h2,
 h3,
